move tests to proper files
